[FancyZones] Auto-zone sibling windows under mouse or last known zone
Updated the "Move newly created windows to their last known zone" code to first check if the window is a "sibling", or, new window of the same process we're currently working on.
- If so, the new window is spawned over its sibling.
- If not, the new window is opened to its last known zone.

I find this behavior more intuitive for my workflows, for example, when:
- Opening a browser window and putting it in Zone 1
- CTRL+N and positioning the sibling in Zone 2
- Clicking back to the browser window in Zone 1
- CTRL+N creates a new window in Zone 1, instead of Zone 2

NOTE: I couldn't get `ZoneSet->GetZoneIndexSetFromWindow()` on the `GetForegroundWindow()` to work, so the current implementation uses GetCursorPos() to figure out the currently active window.
061f0a